Move friendship related tests to separate file
